refactored Engine into SceneManager, makes more logical sense
