+++ src/SceneManager/AbstractScene.ts
