+++ src/SceneManager/Transition.ts
